[no-ci] security: wip: privilege separation

+++ scripts/install.sh
@@ -1,12 +1,13 @@
 #!/bin/bash
-set -e
+set -eu
 
 cp() {
-	command cp --preserve=mode --reflink=auto "$@"
+	command cp -L --preserve=mode --reflink=auto "$@"
 }
 
+# only for installing elf files
 install() {
-	command install -D "$@"
+	command install -s -D "$@"
 }
 
 # $1 - Overlay src directory
@@ -18,7 +19,7 @@ install_overlay() {
 	local i
 	local k
 
-	mapfile -t files < <(find "$src" -type f | sed -E "s|^${src}/?+||g")
+	mapfile -t files < <(find -L "$src" -type f | sed -E "s|^${src}/?+||g")
 
 	for i in "${files[@]}"; do
 		k="$(eval echo "$i")"
@@ -26,9 +27,11 @@ install_overlay() {
 		i="$src/$i"
 		k="$dst/$k"
 
-		mkdir -p "$(dirname "$k")"
+		mkdir -p "${k%/*}"
 
-		cp "$i" "$k"
+		if [ "${k##*/}" != ".gitkeep" ]; then
+			cp "$i" "$k"
+		fi
 	done
 }
 
@@ -39,13 +42,22 @@ install_overlay() {
 # Main overlay
 install_overlay rootdir/ "$DESTDIR"
 
-# Extra Directories
-mkdir -p \
-	"$DESTDIR/$DATA_DIR" \
-	"$DESTDIR/$RUNTIME_DIR" \
-	"$DESTDIR/$BUILD_DIR"
+# Build directory
 
-chmod 777 "$DESTDIR/$BUILD_DIR"
+# TODO: Proper privilege separation
+
+# this must match the uid & gid inside rootdir/usr/lib/sysusers.d/zeus.conf
+chown -v 23248:23248 "$DESTDIR/$BUILD_DIR"
+chmod -v 6770 "$DESTDIR/$BUILD_DIR"
+# 6 -> all files "owned" by same group (directory setgid)
+# 7 -> rwx  for owner
+# 7 -> rwx for group
+# 0 -> ---  for anyone else
+
+# TODO: Check to see if we can somehow use acls in packaging, fakeroot doesnt support them
+#setfacl -d -m user::rwx  "$DESTDIR/$BUILD_DIR"
+#setfacl -d -m group::rwx "$DESTDIR/$BUILD_DIR"
+#setfacl -d -m other::--- "$DESTDIR/$BUILD_DIR"
 
 # Runtime overlays
 for i in runtimes/*; do
@@ -54,9 +66,9 @@ done
 
 # Runtimes
 for i in target/"$BUILD_PROFILE"/librt_*.so; do
-	install -m 644 "./$i" -t "$DESTDIR/$PREFIX/lib/zeus/runtimes"
+	install -m 644 "./$i" -t "$DESTDIR/$RUNTIME_DIR"
 done
 
 # zeus & builder
 install -m 755 "target/$BUILD_PROFILE/zeus" -t "$DESTDIR/$PREFIX/bin"
-install -m 755 "target/$BUILD_PROFILE/builder" -t "$DESTDIR/$PREFIX/share/zeus"
+install -m 755 "target/$BUILD_PROFILE/builder" -t "$DESTDIR/$DATA_DIR"
